fix windows build
