+++ caddy.go
@@ -24,7 +24,6 @@ import (
 	"strconv"
 	"strings"
 	"sync"
-	"syscall"
 	"time"
 
 	"github.com/mholt/caddy/caddyfile"
@@ -722,22 +721,6 @@ func IsLoopback(addr string) bool {
 	return host == "localhost" ||
 		strings.Trim(host, "[]") == "::1" ||
 		strings.HasPrefix(host, "127.")
-}
-
-// checkFdlimit issues a warning if the OS limit for
-// max file descriptors is below a recommended minimum.
-func checkFdlimit() {
-	const min = 8192
-
-	// Warn if ulimit is too low for production sites
-	rlimit := &syscall.Rlimit{}
-	err := syscall.Getrlimit(syscall.RLIMIT_NOFILE, rlimit)
-	if err == nil {
-		if rlimit.Cur < min {
-			fmt.Printf("WARNING: File descriptor limit %d is too low for production servers. "+
-				"At least %d is recommended. Fix with \"ulimit -n %d\".\n", rlimit.Cur, min, min)
-		}
-	}
 }
 
 // Upgrade re-launches the process, preserving the listeners
