+++ rlimit_windows.go
@@ -1,0 +1,6 @@
+package caddy
+
+// checkFdlimit issues a warning if the OS limit for
+// max file descriptors is below a recommended minimum.
+func checkFdlimit() {
+}
